refactor(ui): flatten `event-sources/` directories
- for same reasons as previous coommits
