+++ ui/src/app/event-sources/event-source-editor.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
-
 import {Tabs} from 'argo-ui';
-import {EventSource} from '../../../models';
-import {MetadataEditor} from '../../shared/components/editors/metadata-editor';
-import {ObjectEditor} from '../../shared/components/object-editor/object-editor';
+
+import {EventSource} from '../../models';
+import {MetadataEditor} from '../shared/components/editors/metadata-editor';
+import {ObjectEditor} from '../shared/components/object-editor/object-editor';
 
 export function EventSourceEditor({
     onChange,
