+++ ui/src/app/event-sources/event-source-list.scss
