+++ ui/src/app/event-sources/event-source-list.tsx
@@ -3,26 +3,27 @@ import classNames from 'classnames';
 import * as React from 'react';
 import {useContext, useEffect, useState} from 'react';
 import {Link, RouteComponentProps} from 'react-router-dom';
-import {EventSource, kubernetes} from '../../../../models';
-import {ID} from '../../../event-flow/components/event-flow-details/id';
-import {Utils as EventsUtils} from '../../../sensors/components/utils';
-import {uiUrl} from '../../../shared/base';
-import {ErrorNotice} from '../../../shared/components/error-notice';
-import {Node} from '../../../shared/components/graph/types';
-import {Loading} from '../../../shared/components/loading';
-import {NamespaceFilter} from '../../../shared/components/namespace-filter';
-import {Timestamp} from '../../../shared/components/timestamp';
-import {useCollectEvent} from '../../../shared/components/use-collect-event';
-import {ZeroState} from '../../../shared/components/zero-state';
-import {Context} from '../../../shared/context';
-import {Footnote} from '../../../shared/footnote';
-import {historyUrl} from '../../../shared/history';
-import {services} from '../../../shared/services';
-import {useQueryParams} from '../../../shared/use-query-params';
-import {Utils} from '../../../shared/utils';
-import {EventsPanel} from '../../../workflows/components/events-panel';
-import {EventSourceCreator} from '../event-source-creator';
-import {EventSourceLogsViewer} from '../event-source-log-viewer';
+
+import {EventSource, kubernetes} from '../../models';
+import {ID} from '../event-flow/event-flow-details/id';
+import {Utils as EventsUtils} from '../sensors/components/utils';
+import {uiUrl} from '../shared/base';
+import {ErrorNotice} from '../shared/components/error-notice';
+import {Node} from '../shared/components/graph/types';
+import {Loading} from '../shared/components/loading';
+import {NamespaceFilter} from '../shared/components/namespace-filter';
+import {Timestamp} from '../shared/components/timestamp';
+import {useCollectEvent} from '../shared/components/use-collect-event';
+import {ZeroState} from '../shared/components/zero-state';
+import {Context} from '../shared/context';
+import {Footnote} from '../shared/footnote';
+import {historyUrl} from '../shared/history';
+import {services} from '../shared/services';
+import {useQueryParams} from '../shared/use-query-params';
+import {Utils} from '../shared/utils';
+import {EventsPanel} from '../workflows/components/events-panel';
+import {EventSourceCreator} from './event-source-creator';
+import {EventSourceLogsViewer} from './event-source-log-viewer';
 
 const learnMore = <a href='https://argoproj.github.io/argo-events/concepts/event_source/'>Learn more</a>;
