+++ ui/src/app/event-sources/index.ts
@@ -1,4 +1,4 @@
-import {EventSourceContainer} from './components/event-source-container';
+import {EventSourceContainer} from './event-source-container';
 
 export default {
     component: EventSourceContainer
